# Optimize creation of expression and partial indices