… src/jrd/btr.h
… src/jrd/btr_proto.h
… src/jrd/idx.cpp
… src/jrd/met.epp
… src/jrd/recsrc/IndexTableScan.cpp
… src/jrd/replication/Applier.cpp
… src/jrd/validation.cpp
… src/jrd/validation.h