The idea is to share an expression/condition request between multiple processed records in cases when it's possible (index creation, garbage collection, validation).

Test case:

```
recreate table t (id int) ;
insert into t select row_number() over() from rdb$types, rdb$types, rdb$types;
commit;

set stat on;

create index it0 on t (id);
create index it1 on t (id) where id is not null;
create index it2 on t computed (id+0);
create index it3 on t computed (id+0) where id is not null;

-- v5

SQL> create index it0 on t (id);
Elapsed time = 9.331 sec

SQL> create index it1 on t (id) where id is not null;
Elapsed time = 69.126 sec

SQL> create index it2 on t computed (id+0);
Elapsed time = 72.234 sec

SQL> create index it3 on t computed (id+0) where id is not null;
Elapsed time = 129.571 sec

-- v5 with patch

SQL> create index it0 on t (id);
Elapsed time = 9.369 sec

SQL> create index it1 on t (id) where id is not null;
Elapsed time = 10.920 sec

SQL> create index it2 on t computed (id+0);
Elapsed time = 12.464 sec

SQL> create index it3 on t computed (id+0) where id is not null;
Elapsed time = 14.466 sec

```